Update linting packages and rules + fixes
