+++ package.json
@@ -125,14 +125,25 @@
     "@types/faker": "^4.1.12",
     "@types/jest-axe": "^3.5.0",
     "@types/pretty": "^2.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.0.0",
+    "@typescript-eslint/parser": "^4.0.0",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.2",
+    "babel-eslint": "^10.0.0",
     "enzyme": "^3.11.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.5.0",
     "eslint-config-prettier": "^6.11.0",
+    "eslint-config-react-app": "^6.0.0",
+    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jest": "^24.0.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-no-only-tests": "^2.4.0",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-simple-import-sort": "^5.0.2",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-react-hooks": "^4.0.8",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-testcafe": "^0.2.1",
+    "eslint-plugin-testing-library": "^4.7.0",
     "faker": "^4.1.0",
     "husky": "^7.0.0",
     "jest-axe": "^3.5.0",
