+++ src/constants.ts
@@ -23,6 +23,7 @@ export enum DEV_LANGUAGES {
 
 // ROUTER_LANGUAGES are used to set cimode language without breaking locales related to datepickers etc.
 export const ROUTER_LANGUAGES = { ...SUPPORT_LANGUAGES, ...DEV_LANGUAGES };
+// eslint-disable-next-line @typescript-eslint/no-redeclare
 export type ROUTER_LANGUAGES = SUPPORT_LANGUAGES | DEV_LANGUAGES;
 
 export enum LINKEDEVENTS_CONTENT_TYPE {
