+++ src/domain/event/eventForm/useEventFormSubmitRequests.tsx
@@ -47,4 +47,4 @@ const useUpdateImageRequest = () => {
   };
 };
 
-export { useUpdateImageRequest, useCreateOrUpdateVenueRequest };
+export { useCreateOrUpdateVenueRequest, useUpdateImageRequest };
