+++ src/domain/occurrence/occurrencesFormPart/OccurrencesFormPart.tsx
@@ -1,4 +1,3 @@
-import { ApolloQueryResult } from '@apollo/client';
 import addHours from 'date-fns/addHours';
 import formatDate from 'date-fns/format';
 import isBefore from 'date-fns/isBefore';
@@ -15,7 +14,6 @@ import PlaceSelectorField from '../../../common/components/form/fields/PlaceSele
 import TextInputField from '../../../common/components/form/fields/TextInputField';
 import {
   EventQuery,
-  EventQueryVariables,
   OccurrenceFieldsFragment,
   useAddOccurrenceMutation,
   useDeleteOccurrenceMutation,
