+++ src/utils/updateLocaleParam.ts
@@ -1,10 +1,9 @@
-/**
- * Update locale part of the url
- * @param {string} url
- * @param {string} currentLocale
- * @param {string} value
- * @return {string}
- */
-export default (url: string, currentLocale: string, value: string) => {
+const updateLocaleParam = (
+  url: string,
+  currentLocale: string,
+  value: string
+) => {
   return url.replace(`/${currentLocale}`, `/${value}`);
 };
+
+export default updateLocaleParam;
